# code/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/evaluation-parameters": {
      "get": {
        "operationId": "EvaluationParametersController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all evaluation parameters (optionally filtered by department)",
        "tags": [
          "Evaluation Parameters"
        ]
      },
      "post": {
        "operationId": "EvaluationParametersController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvalParamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create an evaluation parameter (superuser/admin/hod)",
        "tags": [
          "Evaluation Parameters"
        ]
      }
    },
    "/api/evaluation-parameters/status": {
      "get": {
        "operationId": "EvaluationParametersController_getDeptStatus",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get department config status map",
        "tags": [
          "Evaluation Parameters"
        ]
      }
    },
    "/api/evaluation-parameters/dept/{department}": {
      "get": {
        "operationId": "EvaluationParametersController_getDraftsByDept",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get draft parameters for a department",
        "tags": [
          "Evaluation Parameters"
        ]
      }
    },
    "/api/evaluation-parameters/{id}/dept/{department}": {
      "patch": {
        "operationId": "EvaluationParametersController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEvalParamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a parameter by ID and department",
        "tags": [
          "Evaluation Parameters"
        ]
      },
      "delete": {
        "operationId": "EvaluationParametersController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a parameter (superuser/admin only)",
        "tags": [
          "Evaluation Parameters"
        ]
      }
    },
    "/api/evaluation-parameters/dept/{department}/approve": {
      "post": {
        "operationId": "EvaluationParametersController_approve",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Approve dept params if total weight = 100% (superuser/dean only)",
        "tags": [
          "Evaluation Parameters"
        ]
      }
    },
    "/api/evaluation-parameters/dept/{department}/submit": {
      "post": {
        "operationId": "EvaluationParametersController_submit",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Submit dept params for approval (hod only)",
        "tags": [
          "Evaluation Parameters"
        ]
      }
    },
    "/api/evaluation-parameters/dept/{department}/reject": {
      "post": {
        "operationId": "EvaluationParametersController_reject",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reject dept params to request revision (superuser/dean only)",
        "tags": [
          "Evaluation Parameters"
        ]
      }
    },
    "/api/evaluation-parameters/bulk": {
      "post": {
        "operationId": "EvaluationParametersController_bulkImport",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkImportEvalParamsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Bulk import evaluation parameters from JSON array (superuser/admin only)",
        "tags": [
          "Evaluation Parameters"
        ]
      }
    },
...
  },
  "info": {
    "title": "ENDUR API",
    "description": "\n      REST API for the **ENDUR Performance Review & Feedback Management System**.\n\n      ## Role-Based Access Control\n      All protected endpoints require the `X-Role` header with one of:\n      `superuser` | `admin` | `dean` | `hod` | `faculty` | `student`\n\n      Additionally, pass `X-User-Id` and `X-User-Name` headers to attribute audit logs correctly.\n\n      ## Authentication\n      Use `POST /api/auth/login` to verify credentials. The response contains the user object\n      which you should store client-side (session) and use to populate role headers.\n      ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateEvalParamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Teaching Clarity"
          },
          "description": {
            "type": "string",
            "example": "Measures how clearly the faculty explains concepts."
          },
          "category": {
            "type": "string",
            "example": "Pedagogy"
          },
          "weight": {
            "type": "number",
            "example": 25,
            "description": "Weight percentage (0-100)"
          },
          "type": {
            "type": "string",
            "example": "rating",
            "enum": [
              "rating",
              "text"
            ]
          },
          "department": {
            "type": "string",
            "example": "Computer Science",
            "description": "Department this parameter belongs to"
          }
        },
        "required": [
          "name",
          "weight",
          "department"
        ]
      },
      "UpdateEvalParamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "weight": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "BulkImportEvalParamsDto": {
        "type": "object",
        "properties": {
          "params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateEvalParamDto"
            }
          }
        },
        "required": [
          "params"
        ]
      },
...
    }
  }
}